+++ sentry-ruby/lib/sentry/hub.rb
@@ -120,7 +120,13 @@ module Sentry
       event
     end
 
-    def add_breadcrumb(breadcrumb)
+    def add_breadcrumb(breadcrumb, hint: {})
+      if before_breadcrumb = current_client.configuration.before_breadcrumb
+        breadcrumb = before_breadcrumb.call(breadcrumb, hint)
+      end
+
+      return unless breadcrumb
+
       current_scope.add_breadcrumb(breadcrumb)
     end