+++ sentry-ruby/lib/sentry/utils/real_ip.rb
@@ -13,8 +13,8 @@ module Sentry
         "fc00::/7",       # private IPv6 range fc00::/7
         "10.0.0.0/8",     # private IPv4 range 10.x.x.x
         "172.16.0.0/12",  # private IPv4 range 172.16.0.0 .. 172.31.255.255
-        "192.168.0.0/16" # private IPv4 range 192.168.x.x
-      ].map { |proxy| IPAddr.new(proxy) }
+        "192.168.0.0/16", # private IPv4 range 192.168.x.x
+      ]
 
       attr_reader :ip
 
@@ -22,12 +22,14 @@ module Sentry
         remote_addr: nil,
         client_ip: nil,
         real_ip: nil,
-        forwarded_for: nil
+        forwarded_for: nil,
+        trusted_proxies: []
       )
         @remote_addr = remote_addr
         @client_ip = client_ip
         @real_ip = real_ip
         @forwarded_for = forwarded_for
+        @trusted_proxies = (LOCAL_ADDRESSES + Array(trusted_proxies)).map { |proxy| IPAddr.new(proxy) }.uniq
       end
 
       def calculate_ip
@@ -37,12 +39,16 @@ module Sentry
         # Could be a CSV list and/or repeated headers that were concatenated.
         client_ips    = ips_from(@client_ip)
         real_ips      = ips_from(@real_ip)
-        forwarded_ips = ips_from(@forwarded_for)
+
+        # The first address in this list is the original client, followed by
+        # the IPs of successive proxies. We want to search starting from the end
+        # until we find the first proxy that we do not trust.
+        forwarded_ips = ips_from(@forwarded_for).reverse
 
         ips = [client_ips, real_ips, forwarded_ips, remote_addr].flatten.compact
 
         # If every single IP option is in the trusted list, just return REMOTE_ADDR
-        @ip = filter_local_addresses(ips).first || remote_addr
+        @ip = filter_trusted_proxy_addresses(ips).first || remote_addr
       end
 
       protected
@@ -62,8 +68,8 @@ module Sentry
         end
       end
 
-      def filter_local_addresses(ips)
-        ips.reject { |ip| LOCAL_ADDRESSES.any? { |proxy| proxy === ip } }
+      def filter_trusted_proxy_addresses(ips)
+        ips.reject { |ip| @trusted_proxies.any? { |proxy| proxy === ip } }
       end
     end
   end