+++ sentry-ruby/lib/sentry/event.rb
@@ -20,7 +20,7 @@ module Sentry
     MAX_MESSAGE_SIZE_IN_BYTES = 1024 * 8
 
     attr_accessor(*ATTRIBUTES)
-    attr_reader :configuration, :request, :exception, :stacktrace
+    attr_reader :configuration, :request, :exception, :stacktrace, :threads
 
     def initialize(configuration:, integration_meta: nil, message: nil)
       # this needs to go first because some setters rely on configuration
@@ -102,6 +102,7 @@ module Sentry
       data[:stacktrace] = stacktrace.to_hash if stacktrace
       data[:request] = request.to_hash if request
       data[:exception] = exception.to_hash if exception
+      data[:threads] = threads.to_hash if threads
 
       data
     end
@@ -112,6 +113,11 @@ module Sentry
 
     def add_request_interface(env)
       @request = Sentry::RequestInterface.from_rack(env)
+    end
+
+    def add_threads_interface(backtrace: nil, **options)
+      @threads = ThreadsInterface.new(**options)
+      @threads.stacktrace = initialize_stacktrace_interface(backtrace) if backtrace
     end
 
     def add_exception_interface(exc)