+++ sentry-ruby/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add thread_id to Exception interface [#1291](https://github.com/getsentry/sentry-ruby/pull/1291)
 - always convert trusted proxies to string [#1288](https://github.com/getsentry/sentry-ruby/pull/1288)
   - fixes [#1274](https://github.com/getsentry/sentry-ruby/issues/1274)